pyproject.toml
[tool.poetry]
name = "convAnimation"
version = "0.1.0"
description = "animation of convolution"
authors = ["i13abe <i13abemotoshi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.6.0"
matplotlib = "^3.5.2"
manimgl = "^1.6.1"

[tool.poetry.dev-dependencies]
pysen = {extras = ["lint"], version = "^0.10.1"}
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


# pysen version
[tool.pysen]
version = "0.10"

# pysen lint 設定
[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
mypy_preset = "entry"
line_length = 100

# mypy target設定
[[tool.pysen.lint.mypy_targets]]
  paths = ["./conv_animation"]

# lint 対象path設定
[tool.pysen.lint.source]
includes = [
  ".",
]
# 無視するフォルダー
excludes = [
  # "./"
]

# pytest設定
[tool.pytest.ini_options]
testpaths = ["./tests"]